Pause - Game Manager - Health - Score
Created a Pause Menu attached to the key "Escape".
Game Manager attached to the Player Prefab. It holds the selected level variable (for when we have a Main Menu to select levels and characters). It then spawns a Level Manager based on level selected. It also handles UI Elements unrelated to the Player Variables (Pause, and future developments).
Healthbar added and adjustable from the PlayerController
Score UI as well.